% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation Matrix}
\usage{
cor_matrix(data, mapping, colored = FALSE, ...)
}
\description{
Provides a correlation matrix that can be used together with the package GGally.
}
