% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Display Data as correlations}
\usage{
cor_plot(data, mapping, color = I("grey50"), ...)
}
\arguments{
\item{data}{dataframe}
}
\description{
Display Data as correlations
}
