% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_score_mean.R
\name{factor_score_mean}
\alias{factor_score_mean}
\title{Extract factor scores from latent model}
\usage{
factor_score_mean(object, name, save_as_object = TRUE)
}
\description{
# Extract factor scores from latent model
}
