% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tab.R
\name{fit_tab}
\alias{fit_tab}
\title{The length of a string (in characters).}
\usage{
fit_tab(object, measures = c("chisq", "df", "pvalue", "cfi", "tli",
  "rmsea", "srmr"), wrmr = FALSE, as_text = FALSE,
  reliability = FALSE, scaled = FALSE)
}
\arguments{
\item{string}{input character vector}
}
\value{
numeric vector giving number of characters in each element of the
  character vector.  Missing strings have missing length.
}
\description{
The length of a string (in characters).
}
\examples{
str_length(letters)
str_length(c("i", "like", "programming", NA))
}
\seealso{
\code{\link{nchar}} which this function wraps
}
