% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_txt.R
\name{fit_txt}
\alias{fit_txt}
\title{Extract results from fitted lavaan object for display in markdown text}
\usage{
fit_txt(object, estimators = c("chisq", "df", "pvalue", "cfi", "rmsea",
  "rmsea.ci.lower", "rmsea.ci.upper", "srmr"), wrmr = FALSE,
  scaled = FALSE)
}
\arguments{
\item{object}{fitted lavaan object}
}
\description{
Extract results from fitted lavaan object for display in markdown text
}
\examples{
model <- 'y_1 =~ y1 + y2 + y3 + y4
y_2 =~ y5 + y6 + y7 + y8
x =~ x1 + x2 + x3
y_1 ~ a*x
y_2 ~ b*x
'
fit <- lavaan::cfa(model, PoliticalDemocracy, estimator = "WLSMV")
fit_txt(fit)
}
