% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_txt_new.R
\name{fit_txt_new}
\alias{fit_txt_new}
\title{Extract fit indices from fitted lavaan model}
\usage{
fit_txt_new(object, estimators = c("chisq", "df", "pvalue", "cfi",
  "rmsea", "rmsea.ci.lower", "rmsea.ci.upper", "srmr"), robust = FALSE)
}
\arguments{
\item{object}{A fitted lavaan object.}

\item{estimators}{The estimators you want to extract. As recommended by Kline (2016), these are c("chi", "cfi", "rmsea", "rmsea.ll", "rmsea.ul", "srmr"). Important: Enter values exactly as labelled in lavaan.}

\item{robust}{logical. Indicate whether you want to extract the robust fit measures.}
}
\value{
Returns the fitted estimates.
}
\description{
Extract fit indices from fitted lavaan model
}
\examples{
fit_txt(fit)
}
