% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_profile.R
\name{fun_profile}
\alias{fun_profile}
\title{Fundamental Data Accessor for the Profile from \sQuote{twelvedata}
     This API endpoint is available starting with the Grow plan.}
\usage{
fun_profile(sym, exchange = "", country = "", apikey)
}
\arguments{
\item{sym}{(character)  A (single or vector) symbol for Indices. Defaults to unset.}

\item{exchange}{(optional, character)  Exchange. Defaults to unset.}

\item{country}{(optional, character)  An alpha code or country name. Defaults to unset.}

\item{apikey}{(optional character) An API key override, if missing a value cached from
package startup is used. The startup looks for either a file in the per-package config
directory provided by \code{tools::R_user_dir} (for R 4.0.0 or later), or the
\code{TWELVEDATA_API_KEY} variable.}
}
\value{
The requested data is returned as a \code{list} object.
}
\description{
Retrieve Profile Data from \sQuote{twelvedata}.
}
\details{
\code{fun_profile}.
}
\seealso{
\url{https://twelvedata.com/docs}
}
