% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_statistics.R
\name{fun_statistics}
\alias{fun_statistics}
\title{Fundamental Data Accessor for Main Statistics from \sQuote{twelvedata}
     This API endpoint is available starting with the Pro plan.}
\usage{
fun_statistics(sym, as = c("raw"), exchange = "", country = "", apikey)
}
\arguments{
\item{sym}{(character)  A (single or vector) symbol for Indices. Defaults to unset.}

\item{as}{(character) Return format as \dQuote{raw}.}

\item{exchange}{(optional, character)  Exchange. Defaults to unset.}

\item{country}{(optional, character)  An alpha code or country name. Defaults to unset.}

\item{apikey}{(optional character) An API key override, if missing a value cached from
package startup is used. The startup looks for either a file in the per-package config
directory provided by \code{tools::R_user_dir} (for R 4.0.0 or later), or the
\code{TWELVEDATA_API_KEY} variable.}
}
\value{
The requested data is returned as a \code{data.frame} object.
}
\description{
Retrieve Main Fundamental Data Statistics Including Valuation
Metrics and Financials from \sQuote{twelvedata}.
}
\details{
\code{fun_statistics}.
}
\seealso{
\url{https://twelvedata.com/docs}
}
