% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{get_quote}
\alias{get_quote}
\title{Quote Data Accessor for \sQuote{twelvedata}}
\usage{
get_quote(sym, interval = c("1min", "5min", "15min", "30min", "45min", "1h",
  "2h", "4h", "1day", "1week", "1month"), as = c("data.frame", "raw"),
  exchange = "", country = "", volume_time_period = 9,
  type = c(NA_character_, "Stock", "Index", "ETF", "REIT"), dp = 5,
  timezone = NA_character_, apikey)
}
\arguments{
\item{sym}{(character) A (single or vector) symbol understood by the backend as a stock
symbol, foreign exchange pair, or more. See the \sQuote{twelvedata} documentation for
details on what is covered.}

\item{interval}{(character) A valid interval designator ranging form \dQuote{1min} to
\dQuote{1month}. Currently supported are 1, 5, 15, 30 and 45 minutes, 1, 2, 4 hours (using
suffix \sQuote{h}, as well as \dQuote{1day}, \dQuote{1week} and \dQuote{1month}.}

\item{as}{(optional, character) A selector for the desired output format: one of
\dQuote{data.frame} (the default) or or \dQuote{raw}.}

\item{exchange}{(optional, character) A selection of the exchange for which data for
\dQuote{sym} is requested, default value is unset.}

\item{country}{(optional, character) A selection of the country exchange for which data
for \dQuote{sym} is requested, default value is unset.}

\item{volume_time_period}{(optional, numeric) The number of days to use when computing
average volume, default is 9.}

\item{type}{(optional, character) A valid security type selection, if set it must be one of
\dQuote{Stock} (the default), \dQuote{Index}, \dQuote{ETF} or \dQuote{REIT}. Default is
unset via the \code{NA} character value. This field may require the premium subscription.}

\item{dp}{(optional, numeric) The number of decimal places returned on floating point
numbers. The value can be between 0 and 11, with a default value of 5.}

\item{timezone}{(optional, character) The timezone of the returned time stamp. This parameter
is optional. Possible values are \dQuote{Exchange} (the default) to return the
exchange-supplied value, \dQuote{UTC} to use UTC, or a value IANA timezone name such as
\dQuote{America/New_York} (see \code{link{OlsonNames}} to see the values R knows). Note
that the IANA timezone values are case-sensitive.}

\item{apikey}{(optional character) An API key override, if missing a value cached from
package startup is used. The startup looks for either a file in the per-package config
directory provided by \code{tools::R_user_dir} (for R 4.0.0 or later), or the
\code{TWELVEDATA_API_KEY} variable.}
}
\value{
The requested data is returned as a \code{data.frame} object with as many rows
as there were symbols in the request.
}
\description{
Retrieve Securities Quotes Data from \sQuote{twelvedata}
}
\details{
The function has been named \code{get_quote()} to not clash with the base R function
\code{quote}.
}
\seealso{
\url{https://twelvedata.com/docs}
}
\author{
Dirk Eddelbuettel
}
