% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{invtd}
\alias{invtd}
\title{Invert a day number to a date character string}
\usage{
invtd(x)
}
\arguments{
\item{x}{numeric []; representing a date as from \code{\link{td}}}
}
\value{
character []
}
\description{
Invert a day number to a date character string
}
\examples{
invtd(td(28, 2, 1900))

# artificially vectorised with Vectorize
invtd(30000:30004)

}
\seealso{
\code{\link{td}}
}

