% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{monthdays}
\alias{monthdays}
\title{Number of days in a month}
\usage{
monthdays(monthno, yearno = 1999L)
}
\arguments{
\item{monthno}{numeric []; month number}

\item{yearno}{numeric []; year}
}
\value{
numeric []
}
\description{
Number of days in a month
}
\examples{
monthdays(1:12, 2000)

monthdays(13, 2000) # NA

}

