% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_round.R
\name{my_round}
\alias{my_round}
\title{Various Ways to Round Numbers}
\usage{
my_round(x, what)
}
\arguments{
\item{x}{numeric.}

\item{what}{string or numeric.}
}
\description{
Various ways to round numbers. Useful for reproducible manuscripts.
}
