% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_crypto_exchanges.R
\name{ref_crypto_exchanges}
\alias{ref_crypto_exchanges}
\title{Reference Data Accessor for Crypto Exchanges from \sQuote{twelvedata}}
\usage{
ref_crypto_exchanges(as = c("data.frame", "raw"), apikey)
}
\arguments{
\item{as}{(optional, character) A selector for the desired output format: one of
\dQuote{data.frame} (the default) or or \dQuote{raw}.}

\item{apikey}{(optional character) An API key override, if missing a value cached from
package startup is used. The startup looks for either a file in the per-package config
directory provided by \code{tools::R_user_dir} (for R 4.0.0 or later), or the
\code{TWELVEDATA_API_KEY} variable.}
}
\value{
The requested data is returned as a \code{data.frame} object.
}
\description{
Retrieve Reference Data for Crypto Exchanges from \sQuote{twelvedata}
}
\details{
\code{ref_crypto_exchanges}.
}
\seealso{
\url{https://twelvedata.com/docs}
}
