% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Time Series Data Accessor for \sQuote{twelvedata}}
\usage{
time_series(sym, interval = c("1min", "5min", "15min", "30min", "45min", "1h",
  "2h", "4h", "1day", "1week", "1month"), outputsize = 30,
  as = c("data.frame", "xts", "raw"), apikey)
}
\arguments{
\item{sym}{(character) A symbol understood as the backend such a stock symbol, foreign
exchange pair, or more. See the \sQuote{twelvedata} documentation.}

\item{interval}{(character) A valid interval designator ranging form \dQuote{1min} to
\dQuote{1month}. Currently supported are 1, 5, 15, 30 and 45 minutes, 1, 2, 4 hours (using
suffix \sQuote{h}, as well as \dQuote{1day}, \dQuote{1week} and \dQuote{1month}.}

\item{outputsize}{(numeric) The requested number of data points, defaults to 30, and bounded
by 1 and 500.}

\item{as}{(character) A selector for the desired output format: one of \dQuote{data.frame}
(the default), \dQuote{xts} (requiring the package to be installed), or \dQuote{raw}.}

\item{apikey}{(optional character) An API key override, if missing a value cached from
package startup is used. The startup looks for either a file in the per-package config
directory provided by \code{tools::R_user_dir}, or the \code{TWELVEDATA_API_KEY} variable.}
}
\value{
The requested data is returned in the requested format containing columns for
data(time), open, high, low, close, and volume. If the request was unsuccessful,
an error message is returned. The date or datetime column is returned parsed using
\code{anytime} if the package is installed.
}
\description{
Retrieve Time Series Data from \sQuote{twelvedata}
}
\details{
This function access time series data from \sQuote{twelvedata}. It requires an API key
to be registered and to be supplied either in user-config file (TODO: add simple
writer) or an environment variable.

The supported API is richer than what the function currently supports, notably with
respect to \emph{multiple} symbols in one request. We expect to expand the function.
}
\examples{
\dontrun{  # requires API key
data <- time_series("SPY", "5min", 500, "xts")
if (requireNamespace("quantmod", quietly=TRUE))
   chartSeries(data)  # convenient plot function for OHLCV data
}
}
\seealso{
\url{https://twelvedata.com/docs}
}
\author{
Dirk Eddelbuettel
}
