% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{yeardays}
\alias{yeardays}
\title{Number of days in a year}
\usage{
yeardays(yearno)
}
\arguments{
\item{yearno}{numeric []; year}
}
\value{
numeric []
}
\description{
Number of days in a year
}
\examples{
yeardays(1900:1910)

# yes, 1900 wasn't a leap year!

}

