% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffs_txt.R
\name{coeffs_txt}
\alias{coeffs_txt}
\title{Extract text with results}
\usage{
coeffs_txt(
  object,
  label_effect = NULL,
  lhs_effect = NULL,
  rhs_effect = NULL,
  indirect_effect = NULL,
  one_sided = FALSE,
  std_type = "std.all"
)
}
\arguments{
\item{object}{fitted lavaan object}
}
\description{
Extract text with results from fitted lavaan object.
}
