% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tab.R
\name{fit_tab}
\alias{fit_tab}
\title{Extract table of results from fitted lavaan object}
\usage{
fit_tab(object, measures = c("chisq", "df", "pvalue", "cfi", "tli",
  "rmsea", "srmr"), wrmr = FALSE, as_text = FALSE,
  reliability = FALSE, scaled = FALSE)
}
\arguments{
\item{object}{fitted lavaan object}
}
\description{
Extract table of results from fitted lavaan object
}
