% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tab.R
\name{fit_tab}
\alias{fit_tab}
\title{Extract table of results from fitted lavaan object}
\usage{
fit_tab(object, measures = c("chisq", "df", "pvalue", "cfi", "tli",
  "rmsea", "srmr"), wrmr = FALSE, as_text = FALSE,
  reliability = FALSE, scaled = FALSE)
}
\arguments{
\item{object}{fitted lavaan object}
}
\description{
Extract table of results from fitted lavaan object
}
\examples{
model <- 'y_1 =~ y1 + y2 + y3 + y4
y_2 =~ y5 + y6 + y7 + y8
x =~ x1 + x2 + x3
y_1 ~ a*x
y_2 ~ b*x
'
fit <- lavaan::cfa(model, PoliticalDemocracy, estimator = "WLSMV")
fit_tab(fit)
}
