% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_txt_new.R
\name{fit_txt}
\alias{fit_txt}
\title{Extract fit indices from fitted lavaan model}
\usage{
fit_txt(object, estimators = c("chi", "cfi", "rmsea", "rmsea.ll",
  "rmsea.ul", "srmr"), robust = FALSE)
}
\arguments{
\item{object}{A fitted lavaan object.}

\item{estimators}{The estimators you want to extract. As recommended by Kline (2016), these are c("chi", "cfi", "rmsea", "rmsea.ll", "rmsea.ul", "srmr"). Important: Enter values exactly as labelled in lavaan.}

\item{robust}{logical. Indicate whether you want to extract the robust fit measures.}
}
\value{
Returns the fitted estimates.
}
\description{
Extract fit indices from fitted lavaan model
}
\examples{
fit_txt(fit)
}
