% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_txt.R
\name{fit_txt}
\alias{fit_txt}
\title{Extract results from fitted lavaan object for display in markdown text}
\usage{
fit_txt(object, estimators = c("chisq", "df", "pvalue", "cfi", "rmsea",
  "rmsea.ci.lower", "rmsea.ci.upper", "srmr"), wrmr = FALSE,
  scaled = FALSE)
}
\arguments{
\item{object}{fitted lavaan object}
}
\description{
Extract results from fitted lavaan object for display in markdown text
}
